add date and improve invoice

+++ src/main/java/com/billing/billing_system/model/ProductModel/ProductEntity.java
@@ -20,7 +20,7 @@ public class ProductEntity {
     public String productName;
 
     @Column(name = "QUANTITY")
-    public Long quantity;
+    public Integer quantity;
 
     @Column(name = "PRICE")
     public BigDecimal price;
